+++ src/Controllers/BlockTypeController.php
@@ -20,7 +20,9 @@ class BlockTypeController extends Controller {
             $this->model = $this->model->findOrFail($id);
         }
 
+        if(!$request->has('status')) $request->merge(['status' => 0]);
         $this->model->fill($request->except('fields'));
+
         $this->model->application_id = Soda::getApplication()->id;
         if($fields = $request->input('fields')) {
             $this->model->fields()->sync($fields);