+++ src/Controllers/Traits/CrudableTrait.php
@@ -68,6 +68,7 @@ Trait CrudableTrait {
             $this->model = $this->model->findOrFail($id);
         }
 
+        if(!$request->has('status')) $request->merge(['status' => 0]);
         $this->model->fill($request->input());
         $this->model->application_id = Soda::getApplication()->id;
         $this->model->save();