+++ src/Models/Observers/DynamicObserver.php
@@ -2,6 +2,7 @@
 
 namespace Soda\Cms\Models\Observers;
 
+use Schema;
 use Soda\Cms\Models\AbstractDynamicType;
 use Soda\Cms\Models\Block;
 use Soda\Cms\Models\ModelBuilder;
@@ -26,5 +27,15 @@ class DynamicObserver
     public function deleting(AbstractDynamicType $type)
     {
         $type->deleteTable();
+    }
+
+    public function saving(AbstractDynamicType $type)
+    {
+        if ($type->isDirty('identifier')) {
+            $old_table = 'soda_' . $type->getOriginal('identifier');
+            $new_table = 'soda_' . $type->identifier;
+
+            Schema::rename($old_table, $new_table);
+        }
     }
 }